+++ crates/plugin_api/src/bridge.rs
@@ -191,3 +191,25 @@ pub enum WidgetControlType {
         class: String,
     }
 }
+
+
+// handlers
+
+pub type ConfigCallbackFn = Arc<dyn Fn(&str, &str) + Send + Sync>;
+
+pub trait ConfigCallbackFnExt {
+    fn new<F>(f: F) -> Self
+    where
+        F: Fn(&str, &str) + Send + Sync + 'static;
+}
+
+impl ConfigCallbackFnExt for ConfigCallbackFn {
+    fn new<F>(f: F) -> Self
+    where
+        F: Fn(&str, &str) + Send + Sync + 'static,
+    {
+        Arc::new(f)
+    }
+}
+
+